Rename internal bootstrap files and update import paths to use pkg/config
