+++ cmd/bootstrap/config.go
@@ -2,12 +2,12 @@ package bootstrap
 
 import (
 	"log"
-	"neko-acm/internal/conf"
+	"neko-acm/pkg/config"
 )
 
 // 初始化配置
 func initConfig() error {
-	err := conf.InitConfig()
+	err := config.InitConfig()
 	if err != nil {
 		log.Println("初始化配置失败！")
 		return err
