+++ internal/interfaces/http/middlewares/auth.go
@@ -3,8 +3,8 @@ package middlewares
 import (
 	"github.com/gin-gonic/gin"
 	"log"
-	"neko-acm/internal/conf"
 	"neko-acm/internal/model"
+	"neko-acm/pkg/config"
 	"net/http"
 )
 
@@ -14,7 +14,7 @@ func TokenAuth() gin.HandlerFunc {
 		token := c.GetHeader("Authorization")
 
 		// 验证Token
-		if token != "Bearer "+conf.Conf.Server.Token {
+		if token != "Bearer "+config.Conf.Server.Token {
 			log.Println("Token 错误，拒绝请求")
 			c.JSON(http.StatusUnauthorized, model.RespError("Token 错误，拒绝请求", nil))
 			c.Abort()
