+++ internal/service/solution/save.go
@@ -2,7 +2,7 @@ package solution
 
 import (
 	"neko-acm/internal/model"
-	"neko-acm/utils"
+	"neko-acm/pkg/utils"
 	"strconv"
 	"time"
 )
